+++ src/app/app.module.ts
@@ -20,10 +20,8 @@ import {
 
 import { AppComponent } from './app.component';
 import { CoreModule } from './core/core.module';
-import { ClaimsModule } from './claims/claims.module';
 import { AppRoutingModule } from './app-routing.module';
 import { ClaimModule } from './claim/claim.module';
-import { ClaimModule as NewClaimModule } from './newClaim/claim.module';
 import { PayoutsModule } from './payouts/payouts.module';
 import { PaymentAdjustmentModule } from './payment-adjustment/payment-adjustment.module';
 import { PartiesModule } from './parties/parties.module';
@@ -31,6 +29,7 @@ import { PartyModule } from './party/party.module';
 import { DomainModule } from './domain';
 import { RepairingModule } from './repairing/repairing.module';
 import { DepositsModule } from './deposits/deposits.module';
+import { ClaimMgtModule } from './claim-mgt/claim-mgt.module';
 
 @NgModule({
     declarations: [AppComponent],
@@ -45,9 +44,8 @@ import { DepositsModule } from './deposits/deposits.module';
         MatMenuModule,
         MatSidenavModule,
         MatListModule,
-        ClaimsModule,
         ClaimModule,
-        NewClaimModule,
+        ClaimMgtModule,
         PayoutsModule,
         PaymentAdjustmentModule,
         PartiesModule,
