+++ src/app/claim-mgt/claim-mgt.module.ts
@@ -1,0 +1,7 @@
+import { NgModule } from '@angular/core';
+import { ClaimMgtRouting } from './claim-mgt-routing.module';
+
+@NgModule({
+    imports: [ClaimMgtRouting]
+})
+export class ClaimMgtModule {}
