+++ src/app/claim-mgt/claim/index.ts
@@ -1,0 +1,1 @@
+export * from './claim.module';
