+++ src/app/claim-mgt/claims/index.ts
@@ -1,0 +1,1 @@
+export * from './claims.module';
