+++ src/app/claim-mgt/claims/search-form/search-form-value.ts
@@ -1,0 +1,7 @@
+import * as domain from '../../../gen-damsel/domain';
+import { ClaimStatus } from '../../../papi/model/claim-statuses';
+
+export interface SearchFormValue {
+    party_id?: domain.PartyID;
+    statuses?: ClaimStatus[];
+}
