+++ src/app/thrift/claim-management.service.ts
@@ -6,7 +6,6 @@ import * as ClaimManagement from './gen-nodejs/ClaimManagement';
 import { ClaimSearchResponse } from '../gen-damsel/claim_management';
 import { ClaimSearchQuery as ClaimSearchQueryType } from './gen-nodejs/claim_management_types';
 import { KeycloakService } from 'keycloak-angular';
-import { map } from 'rxjs/operators';
 
 @Injectable()
 export class ClaimManagementService extends ThriftService {
@@ -16,9 +15,6 @@ export class ClaimManagementService extends ThriftService {
 
     // `any` because thrift need unions in statuses, not ClaimStatus[] from ClaimSearchQuery.
     // TODO: need converter???
-    searchClaims = (query: any): Observable<ClaimSearchResponse> => {
-        return this.toObservableAction('SearchClaims')(new ClaimSearchQueryType(query)).pipe(
-            map(r => r)
-        );
-    };
+    searchClaims = (query: any): Observable<ClaimSearchResponse> =>
+        this.toObservableAction('SearchClaims')(new ClaimSearchQueryType(query));
 }
